Electromagnetic Teathers - Initial creation
WIP - Emag teathers. Giving engineers stuffs to do.

Cathode: Connects the tether to the grid, produces power based on currently extended cable. Sits on the hull of the ship, accessed VIA console. Can be manually adjusted while EVA with tools
Cable: Extends out from cathode, up to 15 tiles. Have a quality value that is used to determine power generation based on distance and cable quality.
Control: A computer that extends/retracts the tether, and reports how much power is generated.
Tether: A stackable cable that can be added to the cathode object